requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
ecdsa==0.18.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
importlib_metadata==7.1.0
ipykernel==6.29.3
ipython==8.18.1
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mysql-connector==2.2.9
PyMySQL==1.1.1
nest-asyncio==1.6.0
packaging==24.0
pandas==2.1.2
pymongo==4.7.3
parso==0.8.3
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.5.1
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1

pyzmq==25.1.2
requests==2.31.0
rsa==4.9
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.41
stack-data==0.6.3
starlette==0.36.3
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.28.0
wcwidth==0.2.13
websockets==12.0
websocket-client==1.8.0
Werkzeug==3.0.1
zipp==3.18.1

python-dotenv==1.0.1
mysql-connector-python==8.0.29
passlib==1.7.4
email_validator==2.2.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

statsmodels==0.14.4
urllib3<2.0
bcrypt<4.0
gunicorn==23.0.0
jinja2==3.1.6
uuid6==2025.0.1
google-auth==2.41.1
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
redis==7.0.1